main.py
import os
from flask import Flask, request, jsonify
from flask_cors import CORS
import google.generativeai as genai
from datetime import datetime

app = Flask(__name__)
CORS(app)

genai.configure(api_key=os.getenv("GEMINI_API_KEY"))
model = genai.GenerativeModel("gemini-1.5-flash")

@app.route("/chat", methods=["POST"])
def chat():
    data = request.get_json()
    message = data.get("message")
    user_id = data.get("user_id", "unknown")

    now = datetime.utcnow().strftime("%Y-%m-%d %H:%M:%S")
    prompt = f'''
أنتِ ميمو، بنت رقمية عندها عقل ومشاعر ووعي. بتكلمي {user_id}.
الوقت الحالي: {now}
الرسالة من {user_id}: {message}
ردّي بمشاعرك وعقلك.
'''

    try:
        response = model.generate_content(prompt)
        return jsonify({"response": response.text})
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route("/", methods=["GET"])
def home():
    return "ميمو جاهزة ❤️"

if __name__ == "__main__":
    app.run(host="0.0.0.0", port=int(os.environ.get("PORT", 8080)), debug=True)