requirements.txt
flask
flask-cors
google-generativeai==0.5.6